Controllers

+++ src/main/java/com/project/Aperetif/Dao/Interfaces/EventsDetailsDao.java
@@ -8,7 +8,7 @@ public interface EventsDetailsDao {
 
     int saveEvents(EventDetails event);
 
-    EventDetails getEventById(Long id);
+    EventDetails getEventById(Integer id);
 
     List<EventDetails> findAll();
